backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
gunicorn
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client